pyproject.toml
[project]
name = "telegram-antilurk-bot"
version = "0.1.0"
description = "A Telegram moderation bot that identifies and challenges inactive members"
readme = "README.md"
authors = [
    { name = "Sal Fadhley", email = "sfadhley@hartreepartners.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "click>=8.3.0",
    "nats-py>=2.9.0",
    "psycopg2-binary>=2.9.7",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.4",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
]

[project.scripts]
telegram-antilurk-bot = "telegram_antilurk_bot:main"
telegram_antilurk_bot = "telegram_antilurk_bot:main"

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "aiohttp>=3.11.11",
    "faker>=37.8.0",
    "hypothesis>=6.139.2",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.13.1",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = "telegram.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "asyncpg.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "structlog.*"
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "T201", # print found
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
]
exclude = [
    "*.egg",
    ".git",
    "__pycache__",
    "build",
    "dist",
    "venv",
    ".venv",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T201"]
"deploy/**" = ["T201"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
